Avoid serializing during Theory discovery
The TheoryDiscoverer is attempting to serialize each entry in order to verify that it's serializable and thus can get a unique ID created for it.  But this serialization is resulting in a huge cost for tests that have lots of entries, and we don't actually need to do the serializatio, we just need to verify that it's serializable.  This commit changes the implementation to use the underlying CanSerializeObject function, and removes a few unnecessary allocations from it.