Remove lots of delegate allocations from TestMessageVisitor
Each call to OnMessage is allocating up to 35 delegates, as each DoVisit call is constructing a delegate for the provided Visit method group.  This commit modifies the invocations to pass through a 'this' reference so that the compiler can statically cache the necessary delegates and avoid the per-call allocations.

The static DoVisit method is no longer called, but I left it as-is due to it being public and not wanting to change public surface area with performance optimizations.  I instead added a private DoVisit method that's used by the implementation.