Eliminate extra list/array allocation per TestCaseMessage
When constructing a TestCaseMessage, it allocates a new single-element array to pass down to the base classes.  When the TestAssemblyMessage base class receives this array, it makes a defensive copy into a List (which results in the allocation of the list and its underlying array) because the ctor is public and it can't trust the caller to not subsequently mutate the input.  But with TestCaseMessage, we can trust it.  This commit adds a set of internal ctors to the messages so that TestCaseMessage can simply pass its test case down unwrapped to the base class, and the base class can then just construct an array around it, saving the unnecessary list and array defensive copy allocations.