Decrease allocations in SerializationHelper
SerializationHelper is doing a lot of string manipulation.  In one key hot path, it splits a string (resulting in allocations) and then trims whitespace off of each part (resulting in more allocations).  This commit adds a SubstringTrim method that does the trimming as part of the substring to avoid the second round of allocations. The commit also avoids some unnecessary delegate allocations for method groups, as well as some unnecessary intermediate array allocations.