Reduce allocations in TestMethodTestCase
TestMethodTestCase provides a UniqueID, which is created via a heavy computation involving serialization and hashing.  As such, it's lazily computed on first use, but the implementation is resulting in a closure and delegate allocation when the Lazy<T> is constructed.  This commit changes that to instead use optimistic concurrency and to avoid the Lazy, the closure, and the delegate allocations entirely.

The routine to compute the ID is also improved.  It creates a hex string by creating a string for each byte and then joining all of those strings together; this commit updates that with a more efficient implementation that avoids all of the intermediate strings (and StringBuilder inside of Join), instead allocating a single character array.