Prevent a first-chance FileNotFoundException.
When an xunit test is being debugged with first-chance exceptions on we get two exceptions before the test even starts running. These are benign, but given how it takes time for the debugger to stop, show the exception helper and restart, twice, it takes away anywhere between 10s and 50s for each debugging session.

One can of course turn off breaking on FileNotFoundExceptions, but if the test happens to be testing a file-system related feature where we actually look for FileNotFoundExceptions it quickly becomes really impractical.

NetStandard 1.1 doesn't seem to have File.Exists, but even solving this for 4.5.2 and other platforms would be a huge improvement to the debugging story. Probably not worth using reflection to get this for netstandard 1.1.

Would be nice to fix especially because otherwise xunit is clean in terms of first-chance exceptions (some runners add their own of course, but we can deal with them separately).