Emit MethodName as a test-case property in -nunit output
CI post-processing scripts may need the test method name; many tools use nunit-style output.
When the test is a Theory, the method name has to be parsed from test-case name. Much worse,
when xunit is used with TechTalk.SpecFlow, the test-case name is based on the Scenario name;
the method name is completely lost! Emit method name as property MethodName in nunit output.